requirements.txt
python-decouple==3.8
requests==2.32.3
Django==4.2
redis==6.2.0
django-redis==5.4.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-rest-passwordreset==1.4.1
drf-yasg==1.21.10
drf-spectacular==0.28.0
# psycopg2-binary==2.9.7
# mysqlclient==2.1.1
mysql-connector-python==9.0.0
whitenoise==6.6.0
gunicorn==22.0.0
uvicorn==0.29.0
